requirements.txt
langchain>=0.2.0
langchain-community>=0.0.18
openai
faiss-cpu
python-dotenv
tiktoken
unstructured
streamlit
PyPDF2